+++ src/constants/PathConstants.java
@@ -1,0 +1,8 @@
+package constants;
+
+public class PathConstants {
+    public static final String BASE_PATH = System.getProperty("user.dir") + "/portfolio_performance_file/";
+    public static final String FILE_NAME = "Portfolio Performance.xml";
+    public static final String LOGS_PATH = BASE_PATH + "logs/";
+    public static final String SAVE_FILE = LOGS_PATH + "do_not_delete_needed_for_next_import.txt";
+}
