+++ src/enums/SecurityType.java
@@ -1,0 +1,6 @@
+package enums;
+
+public enum SecurityType {
+    ETF,
+    FOND
+}
